chore: component upgrades: qosdk 6.5.1, quarkus: 3.6.5, josdk: 4.6.2
