+++ docs/modules/developer-guide/pages/development-guide.adoc
@@ -172,34 +172,3 @@ Once you make changes, don't forget to xref:testing.adoc[run tests].
 ----
 ./hack/run-maven.sh clean verify
 ----
-
-== Building Images
-
-All images used by this service can be found in the `src/main/images` directory.
-
-[WARNING]
-====
-Please note that in order to build images you need to be in the *root of the repository*.
-====
-
-
-There are available scripts that help with building all images:
-
-- `./hack/build-images-podman.sh`
-
-  Uses local Podman to build all images.
-
-- `./hack/build-images-docker.sh`
-
-  Uses local Docker to build all images.
-
-- `./hack/build-images-minikube.sh`
-
-  Build images inside of the Minikube environment. This is very useful in case of the `local` development environment
-  where you want to test images in a full Kubernetes deployment.
-
-> :arrow_right: Service needs to be built first!
->
-> If you don't use above scripts -- please make sure you build the project before building images to ensure the latest
-> state of the service is included in the image.
-
