+++ pom.xml
@@ -55,8 +55,8 @@
         https://github.com/quarkiverse/quarkus-operator-sdk/blob/main/pom.xml#L15 for the Quarkus
         Operator SDK extension version defined below ("version.quarkus-operator-sdk").
         -->
-        <version.quarkus-platform>3.4.3</version.quarkus-platform>
-        <version.quarkus-operator-sdk>6.4.1</version.quarkus-operator-sdk>
+        <version.quarkus-platform>3.6.5</version.quarkus-platform>
+        <version.quarkus-operator-sdk>6.5.1</version.quarkus-operator-sdk>
 
         <version.cyclonedx>8.0.3</version.cyclonedx>
         <version.lombok>1.18.30</version.lombok>
